# openapi.yaml
openapi: 3.0.3
info:
  title: API de Charadas
  description: |
    API REST para gerenciamento de charadas (perguntas e respostas).
    
    ### Funcionalidades:
    - Autenticação via token JWT
    - CRUD completo de charadas
    - Consulta de charada aleatória
    - Integração com Firebase Firestore

  version: "1.0.0"
  contact:
    name: Guilherme da Costa Silva

servers:
  - url: http://localhost:5000
    description: Servidor local

tags:
  - name: Geral
    description: Informações da API
  - name: Auth
    description: Autenticação
  - name: Charadas
    description: Operações com charadas

paths:
  /:
    get:
      tags: [Geral]
      summary: Informações da API
      description: Retorna informações básicas da API
      responses:
        '200':
          description: Informações retornadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiInfo'

  /login:
    post:
      tags: [Auth]
      summary: Login do administrador
      description: Autentica o usuário e retorna um token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              exemplo:
                value:
                  usuario: admin
                  senha: 123456
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Dados inválidos
        '401':
          description: Usuário ou senha inválidos

  /charadas:
    get:
      tags: [Charadas]
      summary: Listar todas as charadas
      description: Retorna todas as charadas cadastradas no banco
      responses:
        '200':
          description: Lista retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Charada'

    post:
      tags: [Charadas]
      summary: Criar nova charada
      description: Cria uma nova charada (requer autenticação)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharadaInput'
            examples:
              exemplo:
                value:
                  pergunta: "O que é, o que é? Cai em pé e corre deitado?"
                  resposta: "Chuva"
      responses:
        '201':
          description: Charada criada com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Token inválido ou ausente

  /charadas/aleatoria:
    get:
      tags: [Charadas]
      summary: Obter charada aleatória
      description: Retorna uma charada aleatória do banco
      responses:
        '200':
          description: Charada aleatória
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'

  /charadas/{id}:
    parameters:
      - $ref: '#/components/parameters/CharadaId'

    get:
      tags: [Charadas]
      summary: Buscar charada por ID
      description: Retorna uma charada específica pelo ID
      responses:
        '200':
          description: Charada encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'
        '404':
          description: Charada não encontrada

    put:
      tags: [Charadas]
      summary: Atualizar charada (completo)
      description: Atualiza completamente uma charada existente
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharadaInput'
      responses:
        '200':
          description: Charada atualizada com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Não autorizado
        '404':
          description: Charada não encontrada

    patch:
      tags: [Charadas]
      summary: Atualizar charada parcialmente
      description: Atualiza apenas os campos enviados
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharadaUpdate'
      responses:
        '200':
          description: Charada atualizada com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Não autorizado
        '404':
          description: Charada não encontrada

    delete:
      tags: [Charadas]
      summary: Deletar charada
      description: Remove uma charada pelo ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Charada excluída com sucesso
        '401':
          description: Não autorizado
        '404':
          description: Charada não encontrada

components:
  schemas:
    ApiInfo:
      type: object
      properties:
        api:
          type: string
          example: Charadas
        Version:
          type: string
          example: "1.0"
        Author:
          type: string
          example: Guilherme da Costa Silva

    LoginRequest:
      type: object
      required: [usuario, senha]
      properties:
        usuario:
          type: string
          example: admin
        senha:
          type: string
          example: 123456

    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: Login realizado com sucesso!
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    Charada:
      type: object
      properties:
        id:
          type: integer
          example: 1
        pergunta:
          type: string
          example: "O que é, o que é? Tem dentes mas não morde?"
        resposta:
          type: string
          example: "Pente"

    CharadaInput:
      type: object
      required: [pergunta, resposta]
      properties:
        pergunta:
          type: string
        resposta:
          type: string

    CharadaUpdate:
      type: object
      properties:
        pergunta:
          type: string
        resposta:
          type: string

  parameters:
    CharadaId:
      name: id
      in: path
      required: true
      description: ID da charada
      schema:
        type: integer
        example: 1

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT